+++ app/Http/Controllers/CommentController.php
@@ -2,11 +2,13 @@
 
 namespace App\Http\Controllers;
 
+use App\Exceptions\NotAllowedException;
 use App\Http\Requests\StoreCommentRequest;
 use App\Http\Requests\UpdateCommentRequest;
 use App\Http\Resources\CommentResource;
 use App\Models\Comment;
 use Illuminate\Http\JsonResponse;
+use Illuminate\Support\Facades\Gate;
 use Knuckles\Scribe\Attributes\Group;
 use Symfony\Component\HttpFoundation\Response;
 
@@ -20,6 +22,8 @@ class CommentController extends Controller
      */
     public function store(StoreCommentRequest $request): JsonResponse
     {
+        if (Gate::denies('is-user')) return NotAllowedException::notAllowed();
+
         $result = Comment::create($request->validated());
 
         return response()->json(new CommentResource($result->load(['user', 'product'])));
@@ -42,6 +46,8 @@ class CommentController extends Controller
      */
     public function update(UpdateCommentRequest $request, Comment $comment): JsonResponse
     {
+        if (Gate::denies('update-comment')) return NotAllowedException::notAllowed();
+
         $comment->update($request->validated());
 
         return response()->json(new CommentResource($comment->load(['user', 'product'])));
@@ -54,6 +60,8 @@ class CommentController extends Controller
      */
     public function destroy(Comment $comment): JsonResponse
     {
+        if (Gate::denies('delete-comment')) return NotAllowedException::notAllowed();
+
         $comment->delete();
 
         return response()->json(null, Response::HTTP_NO_CONTENT);
